# Conflicts:
#	eng/pipelines/templates/stages/archetype-sdk-client.yml
#	sdk/spring/monitor-tests.yml
#	sdk/spring/pipeline/monitor-supported-version-matrix.json
#	sdk/spring/spring-cloud-azure-appconfiguration-config/pom.xml
#	sdk/spring/spring-cloud-azure-starter-servicebus-jms/pom.xml
#	sdk/spring/tests.yml

+++ sdk/support/azure-resourcemanager-support/src/main/java/com/azure/resourcemanager/support/fluent/models/FileDetailsInner.java
@@ -10,7 +10,9 @@ import com.azure.core.management.SystemData;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.OffsetDateTime;
 
-/** Object that represents File Details resource. */
+/**
+ * Object that represents File Details resource.
+ */
 @Fluent
 public final class FileDetailsInner extends ProxyResource {
     /*
@@ -25,13 +27,15 @@ public final class FileDetailsInner extends ProxyResource {
     @JsonProperty(value = "systemData", access = JsonProperty.Access.WRITE_ONLY)
     private SystemData systemData;
 
-    /** Creates an instance of FileDetailsInner class. */
+    /**
+     * Creates an instance of FileDetailsInner class.
+     */
     public FileDetailsInner() {
     }
 
     /**
      * Get the innerProperties property: Properties of the resource.
-     *
+     * 
      * @return the innerProperties value.
      */
     private FileDetailsProperties innerProperties() {
@@ -40,7 +44,7 @@ public final class FileDetailsInner extends ProxyResource {
 
     /**
      * Get the systemData property: Azure Resource Manager metadata containing createdBy and modifiedBy information.
-     *
+     * 
      * @return the systemData value.
      */
     public SystemData systemData() {
@@ -49,7 +53,7 @@ public final class FileDetailsInner extends ProxyResource {
 
     /**
      * Get the createdOn property: Time in UTC (ISO 8601 format) when file workspace was created.
-     *
+     * 
      * @return the createdOn value.
      */
     public OffsetDateTime createdOn() {
@@ -58,20 +62,20 @@ public final class FileDetailsInner extends ProxyResource {
 
     /**
      * Get the chunkSize property: Size of each chunk.
-     *
+     * 
      * @return the chunkSize value.
      */
-    public Float chunkSize() {
+    public Integer chunkSize() {
         return this.innerProperties() == null ? null : this.innerProperties().chunkSize();
     }
 
     /**
      * Set the chunkSize property: Size of each chunk.
-     *
+     * 
      * @param chunkSize the chunkSize value to set.
      * @return the FileDetailsInner object itself.
      */
-    public FileDetailsInner withChunkSize(Float chunkSize) {
+    public FileDetailsInner withChunkSize(Integer chunkSize) {
         if (this.innerProperties() == null) {
             this.innerProperties = new FileDetailsProperties();
         }
@@ -81,20 +85,20 @@ public final class FileDetailsInner extends ProxyResource {
 
     /**
      * Get the fileSize property: Size of the file to be uploaded.
-     *
+     * 
      * @return the fileSize value.
      */
-    public Float fileSize() {
+    public Integer fileSize() {
         return this.innerProperties() == null ? null : this.innerProperties().fileSize();
     }
 
     /**
      * Set the fileSize property: Size of the file to be uploaded.
-     *
+     * 
      * @param fileSize the fileSize value to set.
      * @return the FileDetailsInner object itself.
      */
-    public FileDetailsInner withFileSize(Float fileSize) {
+    public FileDetailsInner withFileSize(Integer fileSize) {
         if (this.innerProperties() == null) {
             this.innerProperties = new FileDetailsProperties();
         }
@@ -104,20 +108,20 @@ public final class FileDetailsInner extends ProxyResource {
 
     /**
      * Get the numberOfChunks property: Number of chunks to be uploaded.
-     *
+     * 
      * @return the numberOfChunks value.
      */
-    public Float numberOfChunks() {
+    public Integer numberOfChunks() {
         return this.innerProperties() == null ? null : this.innerProperties().numberOfChunks();
     }
 
     /**
      * Set the numberOfChunks property: Number of chunks to be uploaded.
-     *
+     * 
      * @param numberOfChunks the numberOfChunks value to set.
      * @return the FileDetailsInner object itself.
      */
-    public FileDetailsInner withNumberOfChunks(Float numberOfChunks) {
+    public FileDetailsInner withNumberOfChunks(Integer numberOfChunks) {
         if (this.innerProperties() == null) {
             this.innerProperties = new FileDetailsProperties();
         }
@@ -127,7 +131,7 @@ public final class FileDetailsInner extends ProxyResource {
 
     /**
      * Validates the instance.
-     *
+     * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     public void validate() {
